+++ lib/service/config.dart
@@ -601,9 +601,10 @@ class Config with Service, NetworkAuthProvider, NotificationsListener {
 
   // Getters: otherUniversityServices
   String? get assetsUrl => getOtherServicesUrl(key: 'assets_url');
+  String? get deepLinkRedirectUrl => JsonUtils.stringValue(otherUniversityServices['deep_link_redirect_url']);
 
   // Getters: secretKeys
-  String? get coreOrgId => JsonUtils.stringValue(secretCore['org_id']);
+  String? get coreOrgId        => JsonUtils.stringValue(secretCore['org_id']);
 
   // Getters: settings
   int  get refreshTimeout           => JsonUtils.intValue(settings['refreshTimeout'])  ?? 0;
@@ -615,12 +616,6 @@ class Config with Service, NetworkAuthProvider, NotificationsListener {
   // Getters: path keys access
   dynamic pathEntry(String key)     => MapPathKey.entry(content, key);
   String? stringPathEntry(String key, { String? defaults }) => JsonUtils.stringValue(pathEntry(key)) ?? defaults;
-
-  // Getters: other
-  String? get deepLinkRedirectUrl {
-    Uri? assetsUri = StringUtils.isNotEmpty(assetsUrl) ? Uri.tryParse(assetsUrl!) : null;
-    return (assetsUri != null) ? "${assetsUri.scheme}://${assetsUri.host}/html/redirect.html" : null;
-  }
 
   // Getters: Firebase
   String? get firebaseApiKey => JsonUtils.stringValue(firebase['apiKey']);