+++ CHANGELOG.md
@@ -18,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Upgrade to Flutter 3.29.2, upgraded plugin [#4899](https://github.com/rokwire/illinois-app/issues/4899).
 - Updated Events2.eventDetailUrl parameter types.
 - Extended NotificationService APIs [#576](https://github.com/rokwire/app-flutter-plugin/issues/576).
-
+- Read deep links redirect url from separate config entry [#4888](https://github.com/rokwire/app-flutter-plugin/issues/4888).
 ## [1.8.3] - 2025-03-12
 ### Added
 - Created CompactRoundedButton widget [#4872](https://github.com/rokwire/illinois-app/issues/4872).